Add the "intercept" subcommand and a Windows implementation
This uses WinDivert to redirect packets on an interface to userland so that
we can run a Geneva strategy against them. Currently it will look at all
packets on an interface; in the future this could (will?) be restricted to
traffic to and from a proxy.

This is a rough cut; I expect there to be more commits to refine this.