Type assertion failure and buffer overflows should not cause a panic
Real-world testing suggests that not checking a type assertion correctly
causes panics. Oops. In addition to this, if a TCP trigger uses the "load"
keyword with a value that is larger than the payload itself, this should
fail gracefully instead of cause an index-out-of-bounds panic. I'm sure
there are more places where both of these things will need to be checked;
they'll happen in separate commits.