Remove dependency on getlantern/errors and reformat error messages
Using the getlantern/errors package was a good idea at the time, but it has
grown some extra dependencies in the last few months due to its use
elsewhere in the Lantern stack. I want to keep this library lightweight
enough to be able to use it in multiple projects without bringing a lot of
extra baggage with it for little to no benefit.